Add day 5 code

+++ 2023/src/day5.h
@@ -1,2 +1,2 @@
-int day5Part1();
-int day5Part2();
+long day5Part1();
+long day5Part2();
